In need of proper testing, but better.
Raft of changes and updates mostly (though not all) in support of adding an I2C attached LCD to the Arduino to enable the firmware to display on going status information.  The "*_Lite.*" source files are clearly based on other peoples work (attribution maintained in the source files) but required work to either incorporate support for a 20x4 display or to reduce the memory footprint to allow everything to fit into 2 KBytes and leave enough space for the stack.