Tuneable Constants
Include ability to tune "constant" values without the need to recompile the firmware.  Far from all constants are handled (some have serious performance impacts when converted to "in memory" constants), but a good start has been made.  See the 'Q' command in the communications protocol and the files "Constants.h" and "Constants.cpp"