Version 1.3.4
Added new command for setting the whole state of a mobile decoder (speed, direction and all function settings), see the 'W' command.  Have also removed all support for partial emulation of the DCC++ firmware (based on similar Arduino hardware) as this was not a complete emulation and I simply wasn't testing or using it.